+++ packages/db/src/index.ts
@@ -3,6 +3,7 @@ export * from './client';
 export * from './migrate';
 export * from './notes-repository';
 export * from './note-versions-repository';
+export * from './entity-provenance-repository';
 export * from './spaces-repository';
 export * from './search-repository';
 export * from './tokens-repository';